# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Locket Clone API Contract
  description: "Tài liệu đặc tả API chuẩn OpenAPI 3.0 / Swagger UI dành cho Frontend\
    \ (Web & Mobile). Cung cấp đầy đủ các endpoint xác thực (JWT), chia sẻ ảnh (Locket),\
    \ quản lý giao dịch tài chính, kết bạn và lưu trữ ảnh Garage S3."
  contact:
    name: Development Team
    email: dev@locketclone.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Local Development Server
security:
- Bearer Authentication: []
...
paths:
...
  /api/v1/expense/savings-goals/{monthKey}:
    get:
      tags:
      - Expenses & Finance
      summary: Xem mục tiêu tiết kiệm
      description: Lấy mục tiêu tiết kiệm của tháng.
      operationId: getSavingsGoal
      parameters:
      - name: monthKey
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SavingsGoalResponse"
    put:
      tags:
      - Expenses & Finance
      summary: Thiết lập mục tiêu tiết kiệm
      description: Tạo mới hoặc cập nhật mục tiêu tiết kiệm tháng.
      operationId: upsertSavingsGoal
      parameters:
      - name: monthKey
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavingsGoalUpsertRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SavingsGoalResponse"
  /api/v1/expense/budgets/{monthKey}:
    get:
      tags:
      - Expenses & Finance
      summary: Xem ngân sách theo tháng
      description: "Lấy ngân sách đã thiết lập cho tháng cụ thể (dạng YYYYMM, ví dụ\
        \ 202610)."
      operationId: getBudget
      parameters:
      - name: monthKey
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BudgetResponse"
    put:
      tags:
      - Expenses & Finance
      summary: Thiết lập ngân sách tháng
      description: Tạo mới hoặc cập nhật hạn mức ngân sách tháng.
      operationId: upsertBudget
      parameters:
      - name: monthKey
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetUpsertRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BudgetResponse"
...
  /api/v1/expense/categories:
    get:
      tags:
      - Expenses & Finance
      summary: Lấy danh sách danh mục chi tiêu/thu nhập
      description: Lấy danh sách các danh mục hệ thống và danh mục tùy chỉnh của người
        dùng hiện tại.
      operationId: getCategories
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryResponse"
    post:
      tags:
      - Expenses & Finance
      summary: Tạo danh mục mới
      description: Tạo danh mục chi tiêu hoặc thu nhập tùy chỉnh.
      operationId: createCategory
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CategoryResponse"
  /api/v1/auth/verify:
    post:
      tags:
      - Auth
      summary: Xác minh mã OTP
      description: Xác nhận mã OTP 6 chữ số gửi qua email. Nếu thành công trả về JWT
        token và kích hoạt tài khoản. Cho phép thử tối đa 5 lần.
      operationId: verify
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JwtResponse"
  /api/v1/auth/register:
    post:
      tags:
      - Auth
      summary: Đăng ký tài khoản mới
      description: "Gửi thông tin đăng ký (email, username, password). Hệ thống gử\
        i mã OTP xác nhận qua email bất đồng bộ."
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RegisterResponse"
  /api/v1/auth/login:
    post:
      tags:
      - Auth
      summary: Đăng nhập
      description: "Đăng nhập bằng email/username và mật khẩu, trả về Access Token\
        \ JWT."
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JwtResponse"
...
  /api/v1/expense/categories/{categoryId}:
    patch:
      tags:
      - Expenses & Finance
      summary: Cập nhật danh mục
      description: Chỉnh sửa tên hoặc icon của danh mục.
      operationId: updateCategory
      parameters:
      - name: categoryId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CategoryResponse"
...
  /api/v1/expense/summary:
    get:
      tags:
      - Expenses & Finance
      summary: Báo cáo tổng kết chi tiêu tháng
      description: "Tổng hợp tổng chi tiêu, ngân sách còn lại, tỷ lệ sử dụng."
      operationId: getExpenseSummary
      parameters:
      - name: monthKey
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExpenseSummaryResponse"
  /api/v1/expense/summary/yearly:
    get:
      tags:
      - Expenses & Finance
      summary: Báo cáo dòng tiền cả năm
      description: Tổng hợp dòng tiền 12 tháng trong năm (đã tối ưu 1 query gom nhóm).
      operationId: getYearlyCashflowSummary
      parameters:
      - name: year
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/YearlyCashflowSummaryResponse"
  /api/v1/expense/entries:
    get:
      tags:
      - Expenses & Finance
      summary: Lấy danh sách bản ghi giao dịch theo tháng
      description: Lấy danh sách các khoản thu/chi trong tháng với phân trang.
      operationId: getExpenseEntries
      parameters:
      - name: monthKey
        in: query
        required: true
        schema:
          type: string
      - name: type
        in: query
        required: false
        schema:
          type: string
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageExpenseItemResponse"
  /api/v1/expense/entries/by-period:
    get:
      tags:
      - Expenses & Finance
      summary: Lấy giao dịch theo chu kỳ linh hoạt
      description: "Lọc giao dịch theo DAY, WEEK, MONTH, YEAR hoặc CUSTOM."
      operationId: getExpenseEntriesByPeriod
      parameters:
      - name: period
        in: query
        required: true
        schema:
          type: string
          enum:
          - DAY
          - WEEK
          - MONTH
      - name: referenceDate
        in: query
        required: true
        schema:
          type: string
          format: date
      - name: type
        in: query
        required: false
        schema:
          type: string
          default: EXPENSE
          enum:
          - EXPENSE
          - INCOME
          - ALL
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageExpenseItemResponse"
  /api/v1/expense/categories/top:
    get:
      tags:
      - Expenses & Finance
      summary: Top danh mục chi tiêu nhiều nhất
      description: Lấy danh sách top N danh mục có tổng tiền lớn nhất.
      operationId: getTopCategories
      parameters:
      - name: monthKey
        in: query
        required: false
        schema:
          type: string
      - name: year
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: type
        in: query
        required: false
        schema:
          type: string
          default: EXPENSE
          enum:
          - EXPENSE
          - INCOME
          - ALL
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 5
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopCategoryResponse"
  /api/v1/expense/cashflow:
    get:
      tags:
      - Expenses & Finance
      summary: Báo cáo dòng tiền tháng (Cashflow)
      description: "Tổng thu, tổng chi và dòng tiền thuần theo tháng."
      operationId: getCashflowSummary
      parameters:
      - name: monthKey
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CashflowSummaryResponse"
components:
  schemas:
...
    SavingsGoalUpsertRequest:
      type: object
      properties:
        targetAmount:
          type: number
          minimum: 0.01
      required:
      - targetAmount
    SavingsGoalResponse:
      type: object
      properties:
        monthKey:
          type: string
        targetAmount:
          type: number
        currentSaved:
          type: number
        remaining:
          type: number
        progressPct:
          type: integer
          format: int32
        achieved:
          type: boolean
    BudgetUpsertRequest:
      type: object
      properties:
        amountLimit:
          type: number
        alertThresholdPct:
          type: integer
          format: int32
          maximum: 100
          minimum: 1
      required:
      - amountLimit
    BudgetResponse:
      type: object
      properties:
        monthKey:
          type: string
        amountLimit:
          type: number
        alertThresholdPct:
          type: integer
          format: int32
        spent:
          type: number
        remaining:
          type: number
        exceeded:
          type: boolean
...
    CreateCategoryRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 0
        icon:
          type: string
          maxLength: 50
          minLength: 0
        color:
          type: string
          maxLength: 20
          minLength: 0
        transactionType:
          type: string
          enum:
          - INCOME
          - EXPENSE
      required:
      - name
    CategoryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        isDefault:
          type: boolean
        isActive:
          type: boolean
        transactionType:
          type: string
          enum:
          - INCOME
          - EXPENSE
    VerifyOtpRequest:
      type: object
      properties:
        email:
          type: string
        otp:
          type: string
          pattern: "^\\d{6}$"
      required:
      - email
      - otp
    JwtResponse:
      type: object
      properties:
        token:
          type: string
        type:
          type: string
        id:
          type: string
          format: uuid
        email:
          type: string
        username:
          type: string
        isVerified:
          type: boolean
        profileCompleted:
          type: boolean
        displayName:
          type: string
        avatarUrl:
          type: string
        nextStep:
          type: string
          enum:
          - COMPLETE_PROFILE
          - HOME
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
        username:
          type: string
          maxLength: 50
          minLength: 0
        password:
          type: string
          maxLength: 2147483647
          minLength: 6
      required:
      - email
      - password
      - username
    RegisterResponse:
      type: object
      properties:
        message:
          type: string
        email:
          type: string
        nextStep:
          type: string
    LoginRequest:
      type: object
      properties:
        identifier:
          type: string
        password:
          type: string
      required:
      - identifier
      - password
...
    UpdateCategoryRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 0
        icon:
          type: string
          maxLength: 50
          minLength: 0
        color:
          type: string
          maxLength: 20
          minLength: 0
        isActive:
          type: boolean
        transactionType:
          type: string
          enum:
          - INCOME
          - EXPENSE
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
...
    PageableObject:
      type: object
      properties:
        paged:
          type: boolean
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
        unpaged:
          type: boolean
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
    SortObject:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
...
    CategorySpendResponse:
      type: object
      properties:
        categoryId:
          type: string
          format: uuid
        categoryName:
          type: string
        totalAmount:
          type: number
    ExpenseSummaryResponse:
      type: object
      properties:
        monthKey:
          type: string
        totalSpent:
          type: number
        budgetLimit:
          type: number
        remaining:
          type: number
        budgetExceeded:
          type: boolean
        percentUsed:
          type: integer
          format: int32
        byCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategorySpendResponse"
    MonthlyCashflowItemResponse:
      type: object
      properties:
        monthKey:
          type: string
        income:
          type: number
        expense:
          type: number
        net:
          type: number
    YearlyCashflowSummaryResponse:
      type: object
      properties:
        year:
          type: integer
          format: int32
        totalIncome:
          type: number
        totalExpense:
          type: number
        netCashflow:
          type: number
        months:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyCashflowItemResponse"
    ExpenseItemResponse:
      type: object
      properties:
        photoId:
          type: string
          format: uuid
        imageUrl:
          type: string
        thumbnailUrl:
          type: string
        amount:
          type: number
        note:
          type: string
        categoryId:
          type: string
          format: uuid
        categoryName:
          type: string
        transactionType:
          type: string
          enum:
          - INCOME
          - EXPENSE
        takenAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    PageExpenseItemResponse:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/ExpenseItemResponse"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
    TopCategoryResponse:
      type: object
      properties:
        categoryId:
          type: string
          format: uuid
        categoryName:
          type: string
        totalAmount:
          type: number
        percentage:
          type: integer
          format: int32
    CashflowSummaryResponse:
      type: object
      properties:
        monthKey:
          type: string
        totalIncome:
          type: number
        totalExpense:
          type: number
        netCashflow:
          type: number
        budgetLimit:
          type: number
        budgetRemaining:
          type: number
        budgetUsedPct:
          type: integer
          format: int32
        budgetExceeded:
          type: boolean
        incomeByCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategorySpendResponse"
        expenseByCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategorySpendResponse"
  securitySchemes:
    Bearer Authentication:
      type: http
      description: Nhập Access Token (JWT) được trả về từ API /api/v1/auth/login hoặc
        /api/v1/auth/verify
      name: Bearer Authentication
      scheme: bearer
      bearerFormat: JWT